+++ src/states_screens/online/server_selection.cpp
@@ -140,7 +140,8 @@ void ServerSelection::init()
     m_searcher->clearListeners();
     m_searcher->addListener(this);
     m_icon_bank->setScale((float)getHeight() / 15.0f / 128.0f);
-    m_server_list_widget->setIcons(m_icon_bank, (float)getHeight() / 12.0f);
+    m_server_list_widget->setIcons(m_icon_bank,
+                                   int(getHeight() / 12.0f));
     m_sort_desc = false;
     /** Triggers the loading of the server list in the servers manager. */
     refresh(true);
